+++ src/Components/Toolbar/Groups/FiltersCategoriesGroup.js
@@ -2,8 +2,8 @@ import React, { useState } from 'react';
 import PropTypes from 'prop-types';
 import { ToolbarGroup } from '@patternfly/react-core';
 
-import CategoryDropdown from '../ToolbarItems/CategoryDropdown';
-import ToolbarItem from '../ToolbarItems/';
+import CategoryDropdown from './ToolbarItem/CategoryDropdown';
+import ToolbarItem from './ToolbarItem/ToolbarItem';
 import { optionsForCategories } from '../constants';
 
 const FilterCategoriesGroup = ({ filterCategories, filters, setFilters }) => {
