+++ src/Components/Toolbar/Groups/ToolbarItem/helpers.test.js
@@ -8,6 +8,7 @@ describe('Components/Toolbar/helpers/handleCheckboxChips', () => {
     const result = handleCheckboxChips(statusParam, statusesParam);
     expect(result).toEqual(expect.arrayContaining(expected));
   });
+
   it('should accept two empty arrays and return an empty array', () => {
     const statusParam = [];
     const statusesParam = [];
@@ -15,6 +16,7 @@ describe('Components/Toolbar/helpers/handleCheckboxChips', () => {
     const result = handleCheckboxChips(statusParam, statusesParam);
     expect(result).toEqual(expect.arrayContaining(expected));
   });
+
   it('should accept empty status and non-empty stateues and return an empty array', () => {
     const statusParam = [];
     const statusesParam = [{ key: 1, value: 'template_name_0' }];
@@ -22,6 +24,7 @@ describe('Components/Toolbar/helpers/handleCheckboxChips', () => {
     const result = handleCheckboxChips(statusParam, statusesParam);
     expect(result).toEqual(expect.arrayContaining(expected));
   });
+
   it('should trim the statuses down to the value of the key given by status', () => {
     const statusParam = [1];
     const statusesParam = [
@@ -42,6 +45,7 @@ describe('Components/Toolbar/helpers/handleSingleChips', () => {
     const result = handleSingleChips(date, comparator);
     expect(result).toEqual(expected);
   });
+
   it('should accept two nulls and return an empty array', () => {
     const date = null;
     const comparator = null;
@@ -49,6 +53,7 @@ describe('Components/Toolbar/helpers/handleSingleChips', () => {
     const result = handleSingleChips(date, comparator);
     expect(result).toEqual(expected);
   });
+
   it('should accept a null date and return an empty array', () => {
     const date = null;
     const comparator = [{ key: 'id:asc', value: 'ID ascending' }];
@@ -56,6 +61,7 @@ describe('Components/Toolbar/helpers/handleSingleChips', () => {
     const result = handleSingleChips(date, comparator);
     expect(result).toEqual(expected);
   });
+
   it('should accept a valid date param and return a non-empty array', () => {
     const date = 'id:asc';
     const comparator = [
@@ -66,12 +72,14 @@ describe('Components/Toolbar/helpers/handleSingleChips', () => {
     const result = handleSingleChips(date, comparator);
     expect(result).toEqual(expected);
   });
+
   it('should accept a valid date param and a comparator of empty objects and return an empty array, not an array of undefined', () => {
     const date = 'id:asc';
     const comparator = [{}, {}];
     const result = handleSingleChips(date, comparator);
     expect(result).toEqual([]);
   });
+
   it('should accept a valid date param and a null comparator and return an empty array, not an array of undefined', () => {
     const date = 'id:asc';
     const comparator = null;
