+++ src/Components/Toolbar/Groups/ToolbarItem/index.js
@@ -1,0 +1,2 @@
+import ToolbarItem from './ToolbarItem';
+export default ToolbarItem;
