+++ src/Components/Toolbar/index.js
@@ -1,3 +1,2 @@
 import Toolbar from './Toolbar';
-
 export default Toolbar;
